Refactor videodisplay out of video selector
